Add benchmark for DFT
- Unfortunately, with these benchmarks I found that the cv::dft function is not optimized
much by passing in UMats vs Mats. In fact, I didn't get any speed up until I increased
the image size to be nearly 4096x4096. This is an unfortunuate finding. This means that
if I want to show any improvements in the code I will have to use the matchTemplate
function to perform the overlapp and add. It surprises me that there is not a GPU
optimized DFT implementation, or at least, if there is one, it is no better than the CPU
version and mostly worse.

Signed-off-by: Cody W. Eilar <Cody.Eilar@Gmail.com>