Add subregion test
- I added a test to verify that when you get a subregion of an image that
you are actually retrieving a reference and not a copy. The good news is that
when you use the cv::Rect() inside a mat, it returns a reference to the subregion
of the original matrix. This is good news for attempting to speed up the template
matching for very large images.

Signed-off-by: Cody W. Eilar <Cody.Eilar@Gmail.com>