Start work on getting centroids and orientations
- Not complete, but I have implemented some tests at this point that demonstrate
how to get the connected componenets and also the centroids and orientations.
The issue that I am having with this is that I can't seem to get the same answer
as MATLAB's region props for orientation. I am, however getting the same answer
that matlab is getting with the centroid values.

Signed-off-by: Cody W. Eilar <Cody.Eilar@Gmail.com>